+++ src/sentry/issues/formatting/limits.py
@@ -1,9 +1,6 @@
-"""Per-section size limits + presets, mirroring Seer's ``EventFormatLimits``.
-
-A profile pairs a section list with one of these presets. Sections apply the relevant
-caps as they render, so output stays bounded regardless of event size. ``None`` means no
-cap for that section. ``max_frames``/``max_breadcrumbs`` are count caps (Seer hardcodes
-these; we fold them in here so a profile fully controls its output size).
+"""Per-section size limits, mirroring Seer's ``EventFormatLimits``. Sections apply these caps
+as they render, so output stays bounded. ``None`` means no cap; ``max_frames``/
+``max_breadcrumbs`` are count caps.
 """
 
 from __future__ import annotations
@@ -26,15 +23,5 @@ class Limits:
 LIMITS_DEFAULT = Limits(
     max_exceptions_chars=100_000,
     max_stacktrace_chars=20_000,
-    max_spans_chars=5_000,
-)
-
-# tight context, e.g. the seer summary path
-LIMITS_LOW = Limits(
-    max_exceptions_chars=50_000,
-    max_stacktrace_chars=10_000,
-    max_request_chars=2_000,
-    max_breadcrumbs_chars=5_000,
-    max_single_breadcrumb_chars=500,
     max_spans_chars=5_000,
 )
