+++ tests/sentry/issues/formatting/test_limits.py
@@ -1,6 +1,6 @@
 import pytest
 
-from sentry.issues.formatting.limits import LIMITS_DEFAULT, LIMITS_LOW, Limits
+from sentry.issues.formatting.limits import LIMITS_DEFAULT, Limits
 
 
 def test_default_preset_matches_seer() -> None:
@@ -13,21 +13,10 @@ def test_default_preset_matches_seer() -> None:
     assert LIMITS_DEFAULT.max_single_breadcrumb_chars is None
 
 
-def test_low_preset_matches_seer() -> None:
-    assert LIMITS_LOW.max_exceptions_chars == 50_000
-    assert LIMITS_LOW.max_stacktrace_chars == 10_000
-    assert LIMITS_LOW.max_request_chars == 2_000
-    assert LIMITS_LOW.max_breadcrumbs_chars == 5_000
-    assert LIMITS_LOW.max_single_breadcrumb_chars == 500
-    assert LIMITS_LOW.max_spans_chars == 5_000
-
-
 def test_count_caps_default() -> None:
     # folded-in count caps (Seer hardcodes these)
     assert LIMITS_DEFAULT.max_frames == 16
     assert LIMITS_DEFAULT.max_breadcrumbs == 10
-    assert LIMITS_LOW.max_frames == 16
-    assert LIMITS_LOW.max_breadcrumbs == 10
 
 
 def test_limits_are_frozen() -> None:
